tableau de bors + autres fenêtres

+++ src/Controller/DonneeController.php
@@ -6,7 +6,9 @@ use App\Entity\Donnee;
 use App\Entity\Fichier;
 use App\Form\AfficherType;
 use App\Repository\DonneeRepository;
+use App\Repository\FichierRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -17,7 +19,7 @@ class DonneeController extends AbstractController
     /**
      * @Route("/tableauDeBord", name="tableauDeBord", methods={"GET","POST"})
      */
-    public function TableauDeBord(DonneeRepository $repo, Request $request)
+    public function tableauDeBord(DonneeRepository $repo, Request $request, PaginatorInterface $paginator)
     {
         $this->denyAccessUnlessGranted('ROLE_USER');
 
@@ -27,45 +29,38 @@ class DonneeController extends AbstractController
 
         $dataAffiche = [];
         $donneesAffichees = [];
+
+        $idF = 0;
         
         if ($form->isSubmitted() && $form->isValid()) {
             $fichierSelect = $form->getData()['fichier'];
-            $idFichier = $fichierSelect->getId();
-            $this->addFlash("success", "Le fichier choisi a pour ID : $idFichier");
 
             /////// Liste de touts les données qui ont été sélectinné
             $dataAffiche = [];
             for ($i = $fichierSelect->getPremiereDonnee()->getId(); $i <= ($fichierSelect->getPremiereDonnee()->getId())+($fichierSelect->getNbLigne())-2; $i++){
                 $dataAffiche[] = $i;
-                //$dataAffiche[] = $repo->findBy(['id' => $i]);
             }
-            $donneesAffichees = $repo->findBy(['id' => $dataAffiche]); // pour avoir un tableau des differentes données à prendre en compt
-    
-            // for ($i = 144; $i <= 144+12; $i++){
-            //     $dataAffiche[] = $i;
-            // }
-            ///////
-            //dump($dataAffiche);
+            $donneesAffichees = $repo->findBy(['id' => $dataAffiche]); // 
+            $idF = $fichierSelect->getId();
 
-            return $this->render('donnee/tableauDeBord.html.twig', [
-            'form' => $form->createView(),
-            "listeDonnee" => $donneesAffichees
-            
-        ]);
         }
 
         return $this->render('donnee/tableauDeBord.html.twig', [
             'form' => $form->createView(),
-            "listeDonnee" => $donneesAffichees
+            "listeDonnee" => $donneesAffichees,
+            "idF" => $idF
+
         ]);
     }
     
-/////////////////////////////////////////////
     /**
      * @Route("/deleteDonnee/{id}", name="deleteDonnee", methods={"GET"})
      */
     public function deleteDonnee(Donnee $donnee, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+        //voir si cette fonction est vraiment necessaire dans l'etats aqutuel des chose et 
+        // basculer deteData de fichier controller ici
         
         $entityManager->remove($donnee);
         $entityManager->flush();
@@ -73,39 +68,77 @@ class DonneeController extends AbstractController
         return $this->redirectToRoute('tableauDeBord');
     }
 
-    // /**
-    //  * @Route("/tableauDeBord", name="tableauDeBord", methods={"GET"})//,"POST"})
-    //  */
-    // public function TableauDeBord(Fichier $fichier = null,DonneeRepository $repo)//, Request $request)
-    // {
-    //     $this->denyAccessUnlessGranted('ROLE_USER'); // pour gérer les accès, si l'utilisateur n'as pas les droit il sera rediriger vers la page de login / il est aussi possible de gerer les accès au niveau des annotation avec celles des routes
-        
-    //     $donnees = $repo->findAll();
-    //     // return $this->render('donnee/tableauDeBord.html.twig',[
-    //     //     "lesDonnees" => $donnees
-    //     // ]);
-    //     //$dataAffiche = [];
+    /**
+     * @Route("/tableauDeBord/listeDonneesComplete/{idF}", name="listeDonneesComplete", methods={"GET"})
+     */
+    public function listeDonneesComplete(int $idF, DonneeRepository $repo, EntityManagerInterface $entityManager): Response
+    {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+        $fichier = $entityManager->getRepository(Fichier::class)->find($idF);
 
-    //     $form = $this->createForm(AfficherType::class, $fichier);
+        $donneesAffichees = [];
 
-    //     //$form->handleRequest($request);
-        
-    //     if ($form->isSubmitted() && $form->isValid()){
-    //         $idFichier = $fichier->getId();
-    //         $this->addFlash("success", "le fichier choisi a pour ID : $idFichier");
-    //         //return $this->redirectToRoute('about');
-    //     }
+        $dataAffiche = [];
 
-    //     return $this->render('donnee/tableauDeBord.html.twig', [
-    //         'form'=> $form->createView(),
-            
-    //         //"lesDonnees" => $donnees
-    //     ]);
-    // }
+        for ($i = $fichier->getPremiereDonnee()->getId(); $i <= ($fichier->getPremiereDonnee()->getId())+($fichier->getNbLigne())-2; $i++){
+            $dataAffiche[] = $i;
+        }
+
+        $donneesAffichees = $repo->findBy(['id' => $dataAffiche]);
+
+        return $this->render('donnee/listeDonneesComplete.html.twig',[
+            "listeDonnee" => $donneesAffichees,
+            "fichier" => $fichier
+        ]);
+    }
+
+     /**
+     * @Route("/tableauDeBord/graphiques/{idF}", name="graphiques", methods={"GET"})
+     */
+    public function graphiques(int $idF, DonneeRepository $repo, EntityManagerInterface $entityManager): Response
+    {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+        $fichier = $entityManager->getRepository(Fichier::class)->find($idF);
+
+        $donneesAffichees = [];
+
+        $dataAffiche = [];
+
+        for ($i = $fichier->getPremiereDonnee()->getId(); $i <= ($fichier->getPremiereDonnee()->getId())+($fichier->getNbLigne())-2; $i++){
+            $dataAffiche[] = $i;
+        }
+
+        $donneesAffichees = $repo->findBy(['id' => $dataAffiche]);
+
+        return $this->render('donnee/graphiques.html.twig',[
+            "listeDonnee" => $donneesAffichees,
+            "fichier" => $fichier
+        ]);
+    }
 
 
+     /**
+     * @Route("/tableauDeBord/statistiques/{idF}", name="statistiques", methods={"GET"})
+     */
+    public function statistiques(int $idF, DonneeRepository $repo, EntityManagerInterface $entityManager): Response
+    {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+        $fichier = $entityManager->getRepository(Fichier::class)->find($idF);
 
+        $donneesAffichees = [];
 
+        $dataAffiche = [];
 
+        for ($i = $fichier->getPremiereDonnee()->getId(); $i <= ($fichier->getPremiereDonnee()->getId())+($fichier->getNbLigne())-2; $i++){
+            $dataAffiche[] = $i;
+        }
+
+        $donneesAffichees = $repo->findBy(['id' => $dataAffiche]);
+
+        return $this->render('donnee/statistiques.html.twig',[
+            "listeDonnee" => $donneesAffichees,
+            "fichier" => $fichier
+        ]);
+    }
 
 }
